Move SSH feature to own package
