+++ examples/electron/package.json
@@ -77,6 +77,7 @@
     "@theia/process": "1.62.0",
     "@theia/property-view": "1.62.0",
     "@theia/remote": "1.62.0",
+    "@theia/remote-ssh": "1.62.0",
     "@theia/remote-wsl": "1.62.0",
     "@theia/scanoss": "1.62.0",
     "@theia/scm": "1.62.0",
