+++ packages/remote-ssh/README.md
@@ -1,0 +1,31 @@
+<div align='center'>
+
+<br />
+
+<img src='https://raw.githubusercontent.com/eclipse-theia/theia/master/logo/theia.svg?sanitize=true' alt='theia-ext-logo' width='100px' />
+
+<h2>ECLIPSE THEIA - REMOTE SSH EXTENSION</h2>
+
+<hr />
+
+</div>
+
+## Description
+
+This package implements functionality to connect to remote systems using Theia via SSH.
+
+## Additional Information
+
+- [API documentation for `@theia/remote-ssh`](https://eclipse-theia.github.io/theia/docs/next/modules/remote-ssh.html)
+- [Theia - GitHub](https://github.com/eclipse-theia/theia)
+- [Theia - Website](https://theia-ide.org/)
+
+## License
+
+- [Eclipse Public License 2.0](http://www.eclipse.org/legal/epl-2.0/)
+- [一 (Secondary) GNU General Public License, version 2 with the GNU Classpath Exception](https://projects.eclipse.org/license/secondary-gpl-2.0-cp)
+
+## Trademark
+
+"Theia" is a trademark of the Eclipse Foundation
+<https://www.eclipse.org/theia>
