+++ packages/remote-ssh/src/electron-node/remote-ssh-backend-module.ts
@@ -1,0 +1,31 @@
+// *****************************************************************************
+// Copyright (C) 2023 TypeFox and others.
+//
+// This program and the accompanying materials are made available under the
+// terms of the Eclipse Public License v. 2.0 which is available at
+// http://www.eclipse.org/legal/epl-2.0.
+//
+// This Source Code may also be made available under the following Secondary
+// Licenses when the conditions for such availability set forth in the Eclipse
+// Public License v. 2.0 are satisfied: GNU General Public License, version 2
+// with the GNU Classpath Exception which is available at
+// https://www.gnu.org/software/classpath/license.html.
+//
+// SPDX-License-Identifier: EPL-2.0 OR GPL-2.0-only WITH Classpath-exception-2.0
+// *****************************************************************************
+
+import { ContainerModule } from '@theia/core/shared/inversify';
+import { ConnectionContainerModule } from '@theia/core/lib/node/messaging/connection-container-module';
+import { RemoteSSHConnectionProvider, RemoteSSHConnectionProviderPath } from '../electron-common/remote-ssh-connection-provider';
+import { RemoteSSHConnectionProviderImpl } from './remote-ssh-connection-provider';
+import { SSHIdentityFileCollector } from './ssh-identity-file-collector';
+
+export const remoteConnectionModule = ConnectionContainerModule.create(({ bind, bindBackendService }) => {
+    bind(RemoteSSHConnectionProviderImpl).toSelf().inSingletonScope();
+    bind(RemoteSSHConnectionProvider).toService(RemoteSSHConnectionProviderImpl);
+    bindBackendService(RemoteSSHConnectionProviderPath, RemoteSSHConnectionProvider);
+});
+
+export default new ContainerModule((bind, _unbind, _isBound, rebind) => {
+    bind(SSHIdentityFileCollector).toSelf().inSingletonScope();
+});
