+++ packages/remote-ssh/src/electron-node/ssh-identity-file-collector.ts
